rta.py: avoid race for RTA JSON config file
During RTA configuration a file is created with the content of the
workload to run. When running LISA in parallel in the same source tree,
a race appears as the configuration file has the same path for all
instances of LISA. Replace that deterministic path by a temporary file
with a unique name. This file is deleted once pushed to the target.

Co-authored-by: Douglas Raillard <douglas.raillard@arm.com>

Signed-off-by: Elieva Pignat <Elieva.Pignat@arm.com>
Signed-off-by: Douglas Raillard <douglas.raillard@arm.com>