exekall: Add support for sequenced operators
Sequenced operators allow the customization module to ask for a given list
operators to never be used interleaved. If the least is [a, b], that means that
values of "a" will stopped being used before values of "b" start being used.

Obviously, a given expression can only use one of the operator in that list,
otherwise it becomes impossible to satisfy the sequencing requirement for other
expressions referencing the same operators. If that happens, an exception is
raised with the list of types returned by the offending operators.