lisa.conf: Allow not deepcopying values upon lookup
Allow the KeyDesc to specify whether a deepcopy of the value should be done upon
lookup. Most of the time, a deepcopy is a good safety measure to ensure the user
will not mutate the value it's given. But sometimes, the object is just too big
to afford a copy.