lisa.analysis.idle: Simplify df_cluster_idle_state_residency()
Rewrite df_cluster_idle_state_residency() to compute the idle state of each CPU
of the cluster at any point in time, and then take the shallowest idle state of
each of them.

Once we have the cluster state, we can compute the time it stayed each time it
enters an idle, and sum it per state.