lisa.utils: Add mp_spawn_pool()
FEATURE

Allow fast creation of a multiprocessing.Pool() using the spawn method,
skipping __main__ import in the new process. If the functions dispatched
to the pool are in a module that is a fast to import, this will provide
a safe replacement for a Pool() created with the default "fork" method,
which is unsafe when threads are present in the application (in Python
or in extension modules).