lisa.trace: Fix FtraceConf merging
FIX

Since the events-namespaces applies to an entire FtraceConf, we need to
"resolve" the namespaces before being able to merge two FtraceConf. This
means the events key need to be able to represent the namespace
information one way or another.

In order to preserve the namespace priority information, make
OrTraceEventChecker order-aware so that "lisa_foo or foo" encodes
accurately the order in which we want to try the events.

Then make TraceEventCheckerBase.check_events() return the spanning set
of events for that checker. This allows easily selecting the events with
the correct priority system.