tools/trace-parser: Relax kallsyms filter
FIX

Only filter out symbols containing "$" in them, as they are probably arm
mapping symbols that bring nothing when trying to do a symbolic
resolution of an address.

We previously filtered out any symbol that is not an identifier, but
that ended up filtering out symbols created by optimization passes, such
as "getname_flags.part.0", leading to the wrong "putname" function to be
detected.